deleted camera picker controller
