+++ HomeTabBarController.swift
@@ -9,12 +9,15 @@
 import UIKit
 import Firebase
 import AFNetworking
+import AVFoundation
+import MobileCoreServices
+import Photos
 
 
-
-class HomeTabBarController: UITabBarController, UITabBarControllerDelegate, UINavigationControllerDelegate {
+class HomeTabBarController: UITabBarController, UITabBarControllerDelegate, UINavigationControllerDelegate, UIImagePickerControllerDelegate {
     
-    
+    var movieCount = Int()
+
     var word: Word?
     
     override func viewDidLoad() {
@@ -54,7 +57,13 @@ class HomeTabBarController: UITabBarController, UITabBarControllerDelegate, UINa
         // Set up the Tab Bar Controller to have four tabs
         
         self.viewControllers = [vc1, vc2, vc3, vc4, vc5]
+    
         
+        FirebaseClient.sharedInstance.fetchMoviePosts { (urlArray) in
+            self.movieCount = (urlArray?.count)!
+        }
+        
+
         setupMiddleButton()
 
         
@@ -90,12 +99,40 @@ class HomeTabBarController: UITabBarController, UITabBarControllerDelegate, UINa
     
     func presentCameraPicker()
     {
-        let storyboard = UIStoryboard(name: "Malcolm.Main", bundle: nil)
-        let nxtNVC = storyboard.instantiateViewController(withIdentifier: "CameraPickerController") as! CameraPickerController
-        self.present( nxtNVC, animated: true, completion: nil)
+        let imagePicker = UIImagePickerController()
+        imagePicker.delegate = self
+        imagePicker.sourceType = .camera
+        imagePicker.videoQuality = UIImagePickerControllerQualityType.typeIFrame960x540//typeIFrame1280x720
+        imagePicker.mediaTypes = [kUTTypeMovie as NSString as String]
+        
+        present(imagePicker, animated: true, completion:nil)
     }
  
     
+    func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [String : Any]) {
+        
+        
+        let url = info[UIImagePickerControllerMediaURL] as! URL
+        
+        
+        self.dismiss(animated: true, completion: nil)
+
+        FirebaseClient.sharedInstance.createNewVideoObject(url: url, movieCount: self.movieCount, complete: {
+            // Dissmissing the camera after successfully upload thus use complete handle
+            // Add HUD while loading
+            
+        })
+        
+        PHPhotoLibrary.shared().performChanges({
+            PHAssetChangeRequest.creationRequestForAssetFromVideo(atFileURL: url )
+            
+        }, completionHandler:nil)
+        
+    }
+    
+    func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
+        dismiss(animated: true, completion:nil)
+    }
