+++ Wordie/Info.plist
@@ -16,8 +16,25 @@
 	<string>APPL</string>
 	<key>CFBundleShortVersionString</key>
 	<string>1.0</string>
+	<key>CFBundleURLTypes</key>
+	<array>
+		<dict>
+			<key>CFBundleURLSchemes</key>
+			<array>
+				<string>fb1828114770805637</string>
+			</array>
+		</dict>
+	</array>
 	<key>CFBundleVersion</key>
 	<string>1</string>
+	<key>FacebookAppID</key>
+	<string>1828114770805637</string>
+	<key>FacebookDisplayName</key>
+	<string>Wordie</string>
+	<key>LSApplicationQueriesSchemes</key>
+	<array>
+		<string>fbauth2</string>
+	</array>
 	<key>LSRequiresIPhoneOS</key>
 	<true/>
 	<key>NSAppTransportSecurity</key>
@@ -33,23 +50,6 @@
 	<string>Smile!</string>
 	<key>UILaunchStoryboardName</key>
 	<string>LaunchScreen</string>
-    <key>CFBundleURLTypes</key>
-    <array>
-        <dict>
-            <key>CFBundleURLSchemes</key>
-            <array>
-                <string>fb1828114770805637</string>
-            </array>
-        </dict>
-    </array>
-    <key>FacebookAppID</key>
-    <string>1828114770805637</string>
-    <key>FacebookDisplayName</key>
-    <string>Wordie</string>
-    <key>LSApplicationQueriesSchemes</key>
-    <array>
-        <string>fbauth2</string>
-    </array>
 	<key>UIMainStoryboardFile</key>
 	<string>Malcolm.Main</string>
 	<key>UIRequiredDeviceCapabilities</key>
